Conflicts:
	src/bldump.c
	src/protocol/protocol.o

+++ src/util.h
@@ -3,12 +3,11 @@
 #define __UTIL_H
 
 #include <stdint.h>
-#include "logevent.h"
 #define getUint32(ev) (*(uint32_t*)(ev))
 #define getUint64(ev) (*(uint64_t*)(ev))
 #define getUint8(ev) (*(uint8_t*)(ev))
 #define getUint16(ev) (*(uint16_t*)(ev))
-#define getInt32(ev) (*(int32_t*)(ev))
+#define getInt32(ev) ((int32_t)(*(uint32_t*)(ev)))
 #define getInt64(ev) (*(int64_t*)(ev))
 #define getInt8(ev) (*(int8_t*)(ev))
 #define getInt16(ev) (*(int16_t*)(ev))
